+++ Msr.Services/Procedures/ViewModels/GetStepEditDataViewModel.cs
@@ -108,7 +108,7 @@ namespace Msr.Services.Procedures.ViewModels
         public List<string> ReferenceFiles { get; set; }
         public decimal? ReplacementCost { get; set; }
         public decimal? Utilization { get; set; }
-        public decimal? UsefulLife { get; set; }
+        public int? UsefulLife { get; set; }
         public decimal? EquipExpensePerMinute { get; set; }
         public decimal? AnnualRM { get; set; }
         public decimal? RMPerMinute { get; set; }
